# Labs/jaxrs-server 2/src/gen/java/io/swagger/model/TapeInput.java
/*
 * Videotapes Galore API
 * Simple user and video tape storing and management system. Has functionality to 'borrow' video tapes to users via relations between the two tapes. Has additional functionality of managing and storing user rating of video tapes and user recommendations of video tapes. API designed by Edda Steinunn Rúnarsdóttir for the assigment Project #1 in the course T-302-HÖNN for Reykjavik University.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


package io.swagger.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import java.util.Date;
import javax.validation.constraints.*;

/**
 * TapeInput
 */
@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaJerseyServerCodegen", date = "2018-09-16T17:01:38.449Z")
public class TapeInput   {
  @JsonProperty("title")
  private String title = null;

  @JsonProperty("director")
  private String director = null;

  @JsonProperty("type")
  private String type = null;

  @JsonProperty("releaseDate")
  private Date releaseDate = null;

  @JsonProperty("eidr")
  private String eidr = null;

  public TapeInput title(String title) {
    this.title = title;
    return this;
  }

  /**
   * Get title
   * @return title
   **/
  @JsonProperty("title")
  @ApiModelProperty(value = "")
  public String getTitle() {
    return title;
  }

  public void setTitle(String title) {
    this.title = title;
  }

  public TapeInput director(String director) {
    this.director = director;
    return this;
  }

  /**
   * Get director
   * @return director
   **/
  @JsonProperty("director")
  @ApiModelProperty(value = "")
  public String getDirector() {
    return director;
  }

  public void setDirector(String director) {
    this.director = director;
  }

  public TapeInput type(String type) {
    this.type = type;
    return this;
  }

  /**
   * Get type
   * @return type
   **/
  @JsonProperty("type")
  @ApiModelProperty(value = "")
  public String getType() {
    return type;
  }

  public void setType(String type) {
    this.type = type;
  }

  public TapeInput releaseDate(Date releaseDate) {
    this.releaseDate = releaseDate;
    return this;
  }

  /**
   * Get releaseDate
   * @return releaseDate
   **/
  @JsonProperty("releaseDate")
  @ApiModelProperty(value = "")
  public Date getReleaseDate() {
    return releaseDate;
  }

  public void setReleaseDate(Date releaseDate) {
    this.releaseDate = releaseDate;
  }

  public TapeInput eidr(String eidr) {
    this.eidr = eidr;
    return this;
  }

  /**
   * Get eidr
   * @return eidr
   **/
  @JsonProperty("eidr")
  @ApiModelProperty(value = "")
  public String getEidr() {
    return eidr;
  }

  public void setEidr(String eidr) {
    this.eidr = eidr;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    TapeInput tapeInput = (TapeInput) o;
    return Objects.equals(this.title, tapeInput.title) &&
        Objects.equals(this.director, tapeInput.director) &&
        Objects.equals(this.type, tapeInput.type) &&
        Objects.equals(this.releaseDate, tapeInput.releaseDate) &&
        Objects.equals(this.eidr, tapeInput.eidr);
  }

  @Override
  public int hashCode() {
    return Objects.hash(title, director, type, releaseDate, eidr);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class TapeInput {\n");
    
    sb.append("    title: ").append(toIndentedString(title)).append("\n");
    sb.append("    director: ").append(toIndentedString(director)).append("\n");
    sb.append("    type: ").append(toIndentedString(type)).append("\n");
    sb.append("    releaseDate: ").append(toIndentedString(releaseDate)).append("\n");
    sb.append("    eidr: ").append(toIndentedString(eidr)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}